+++ patient-search/api/src/main/java/gov/cdc/nbs/controller/OrganizationController.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Controller;
 
 import gov.cdc.nbs.entity.odse.Organization;
 import gov.cdc.nbs.graphql.GraphQLPage;
-import gov.cdc.nbs.graphql.searchFilter.OrganizationFilter;
+import gov.cdc.nbs.graphql.filter.OrganizationFilter;
 import gov.cdc.nbs.service.OrganizationService;
 import lombok.AllArgsConstructor;
