+++ patient-search/api/src/main/java/gov/cdc/nbs/controller/PlaceController.java
@@ -2,7 +2,7 @@ package gov.cdc.nbs.controller;
 
 import gov.cdc.nbs.entity.odse.Place;
 import gov.cdc.nbs.graphql.GraphQLPage;
-import gov.cdc.nbs.graphql.searchFilter.PlaceFilter;
+import gov.cdc.nbs.graphql.filter.PlaceFilter;
 import gov.cdc.nbs.service.PlaceService;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
