+++ patient-search/api/src/main/java/gov/cdc/nbs/graphql/filter/LabReportFilter.java
@@ -1,4 +1,4 @@
-package gov.cdc.nbs.graphql.searchFilter;
+package gov.cdc.nbs.graphql.filter;
 
 import java.time.Instant;
 import java.util.List;
