+++ patient-search/api/src/test/java/gov/cdc/nbs/InvestigationSearchSteps.java
@@ -17,8 +17,8 @@ import org.springframework.transaction.annotation.Transactional;
 import gov.cdc.nbs.controller.EventController;
 import gov.cdc.nbs.entity.elasticsearch.Investigation;
 import gov.cdc.nbs.entity.enums.PregnancyStatus;
-import gov.cdc.nbs.graphql.searchFilter.InvestigationFilter;
-import gov.cdc.nbs.graphql.searchFilter.InvestigationFilter.IdType;
+import gov.cdc.nbs.graphql.filter.InvestigationFilter;
+import gov.cdc.nbs.graphql.filter.InvestigationFilter.IdType;
 import gov.cdc.nbs.repository.JurisdictionCodeRepository;
 import gov.cdc.nbs.repository.elasticsearch.InvestigationRepository;
 import gov.cdc.nbs.support.EventMother;
