+++ patient-search/api/src/test/java/gov/cdc/nbs/LabReportSearchSteps.java
@@ -19,16 +19,16 @@ import org.springframework.transaction.annotation.Transactional;
 import gov.cdc.nbs.controller.EventController;
 import gov.cdc.nbs.entity.elasticsearch.LabReport;
 import gov.cdc.nbs.entity.enums.PregnancyStatus;
-import gov.cdc.nbs.graphql.searchFilter.LabReportFilter;
-import gov.cdc.nbs.graphql.searchFilter.LabReportFilter.EntryMethod;
-import gov.cdc.nbs.graphql.searchFilter.LabReportFilter.EventStatus;
-import gov.cdc.nbs.graphql.searchFilter.LabReportFilter.LabReportDateType;
-import gov.cdc.nbs.graphql.searchFilter.LabReportFilter.LabReportProviderSearch;
-import gov.cdc.nbs.graphql.searchFilter.LabReportFilter.LaboratoryEventDateSearch;
-import gov.cdc.nbs.graphql.searchFilter.LabReportFilter.LaboratoryEventIdType;
-import gov.cdc.nbs.graphql.searchFilter.LabReportFilter.ProcessingStatus;
-import gov.cdc.nbs.graphql.searchFilter.LabReportFilter.ProviderType;
-import gov.cdc.nbs.graphql.searchFilter.LabReportFilter.UserType;
+import gov.cdc.nbs.graphql.filter.LabReportFilter;
+import gov.cdc.nbs.graphql.filter.LabReportFilter.EntryMethod;
+import gov.cdc.nbs.graphql.filter.LabReportFilter.EventStatus;
+import gov.cdc.nbs.graphql.filter.LabReportFilter.LabReportDateType;
+import gov.cdc.nbs.graphql.filter.LabReportFilter.LabReportProviderSearch;
+import gov.cdc.nbs.graphql.filter.LabReportFilter.LaboratoryEventDateSearch;
+import gov.cdc.nbs.graphql.filter.LabReportFilter.LaboratoryEventIdType;
+import gov.cdc.nbs.graphql.filter.LabReportFilter.ProcessingStatus;
+import gov.cdc.nbs.graphql.filter.LabReportFilter.ProviderType;
+import gov.cdc.nbs.graphql.filter.LabReportFilter.UserType;
 import gov.cdc.nbs.repository.JurisdictionCodeRepository;
 import gov.cdc.nbs.repository.elasticsearch.LabReportRepository;
 import gov.cdc.nbs.support.EventMother;
