+++ patient-search/api/src/test/java/gov/cdc/nbs/PatientSearchSteps.java
@@ -20,8 +20,8 @@ import org.springframework.transaction.annotation.Transactional;
 import gov.cdc.nbs.controller.PatientController;
 import gov.cdc.nbs.entity.odse.Person;
 import gov.cdc.nbs.graphql.GraphQLPage;
-import gov.cdc.nbs.graphql.searchFilter.PatientFilter;
-import gov.cdc.nbs.graphql.searchFilter.PatientFilter.Identification;
+import gov.cdc.nbs.graphql.filter.PatientFilter;
+import gov.cdc.nbs.graphql.filter.PatientFilter.Identification;
 import gov.cdc.nbs.repository.PersonRepository;
 import gov.cdc.nbs.repository.PostalLocatorRepository;
 import gov.cdc.nbs.repository.TeleLocatorRepository;
