+++ patient-search/api/src/test/java/gov/cdc/nbs/PermissionSteps.java
@@ -27,10 +27,10 @@ import gov.cdc.nbs.config.security.SecurityProperties;
 import gov.cdc.nbs.controller.EventController;
 import gov.cdc.nbs.controller.PatientController;
 import gov.cdc.nbs.graphql.GraphQLPage;
-import gov.cdc.nbs.graphql.searchFilter.InvestigationFilter;
-import gov.cdc.nbs.graphql.searchFilter.LabReportFilter;
-import gov.cdc.nbs.graphql.searchFilter.OrganizationFilter;
-import gov.cdc.nbs.graphql.searchFilter.PatientFilter;
+import gov.cdc.nbs.graphql.filter.InvestigationFilter;
+import gov.cdc.nbs.graphql.filter.LabReportFilter;
+import gov.cdc.nbs.graphql.filter.OrganizationFilter;
+import gov.cdc.nbs.graphql.filter.PatientFilter;
 import gov.cdc.nbs.repository.ProgramAreaCodeRepository;
 import io.cucumber.java.Before;
 import io.cucumber.java.en.Given;
